+++ test/server/config/application.rb
@@ -5,8 +5,8 @@ require "rails"
 require "action_controller/railtie"
 require "action_cable/engine"
 
-require "yrb_lite"
-require "yrb_lite/action_cable" # YrbLite::ActionCable::Sync (companion gem)
+require "y"
+require "y/action_cable" # Y::ActionCable::Sync (companion gem)
 require_relative "../lib/file_store"
 
 module TestServer